Make main layout working with MUI  v5 and TS
