+++ site/package.json
@@ -3,7 +3,15 @@
     "version": "0.1.0",
     "private": true,
     "dependencies": {
+        "@emotion/react": "^11.11.0",
+        "@emotion/styled": "^11.11.0",
+        "@mui/material": "^5.13.2",
         "fs-watch-exec": "^1.1.2",
+        "react": "18.2.0",
+        "react-dom": "18.2.0",
+        "react-router": "^6.11.2",
+        "react-router-dom": "^6.11.2",
+        "recharts": "^2.6.2",
         "serve": "^14.2.0"
     },
     "scripts": {
@@ -13,8 +21,10 @@
         "dev": "vite --port 3000"
     },
     "devDependencies": {
-        "@types/react": "^18.2.7",
+        "@types/material-ui": "^0.21.12",
+        "@types/react": "^17.0.59",
         "@types/react-dom": "^18.2.4",
+        "@vitejs/plugin-react": "^4.0.0",
         "typescript": "^5.0.4",
         "vite": "^4.3.8"
     }
