+++ site/src/components/home/MainApp.tsx
@@ -1,0 +1,10 @@
+import { AppPannel } from "./AppPannel";
+import { CssBaseline } from "@mui/material";
+import { ThemeSwitcherProvider } from "./ThemeSwitcherProvider";
+
+export function MainApp() {
+    return (<ThemeSwitcherProvider>
+        <CssBaseline />
+        <AppPannel />
+    </ThemeSwitcherProvider>);
+}
