+++ site/src/main.tsx
@@ -1,7 +1,10 @@
+import { createRoot } from "react-dom/client";
+import { MainApp } from "./components/home/MainApp";
+import { StrictMode } from "react";
+
 const rootNode = document.getElementById("root");
 if (rootNode == null) {
         throw new Error("Could not find root element");
 }
 
-// Use React and ReactDOM from UMD bundle instead of from node_modules
-ReactDOM.createRoot(rootNode).render(<React.StrictMode>TEST</React.StrictMode>);
+createRoot(rootNode).render(<StrictMode><MainApp /></StrictMode>);
