+++ site/vite.config.ts
@@ -1,0 +1,11 @@
+// default vite config
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react'
+
+export default defineConfig({
+    plugins: [react()],
+    server: {
+        port: 3000,
+    },
+    publicDir: './local'
+});
